refactor: reorganize history-related classes into AdminHistory package
