+++ TESTS/e2eTests/src/test/java/e2e/AdminHistory/models/HistorySortField.java
@@ -1,8 +1,4 @@
-package e2e.models;
-
-import java.util.ArrayList;
-import java.util.Comparator;
-import java.util.List;
+package e2e.AdminHistory.models;
 
 public enum HistorySortField {
     SHECULED_TIME,
