+++ TESTS/e2eTests/src/test/java/e2e/AdminHistory/models/HistorySorter.java
@@ -1,6 +1,5 @@
-package e2e.models;
+package e2e.AdminHistory.models;
 
-import java.text.Format;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
