+++ TESTS/e2eTests/src/test/java/e2e/AdminHistory/pages/HomePageAdmin.java
@@ -1,14 +1,11 @@
-package e2e.pages;
+package e2e.AdminHistory.pages;
 
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
 import org.openqa.selenium.support.PageFactory;
 import org.openqa.selenium.support.ui.ExpectedConditions;
-import org.openqa.selenium.support.ui.Select;
 import org.openqa.selenium.support.ui.WebDriverWait;
-
-import java.nio.channels.Selector;
 
 public class HomePageAdmin {
     private WebDriver driver;
