+++ TESTS/e2eTests/src/test/java/e2e/AdminHistory/pages/HomeUnregistered.java
@@ -1,4 +1,4 @@
-package e2e.pages;
+package e2e.AdminHistory.pages;
 
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
