+++ TESTS/e2eTests/src/test/java/e2e/AdminHistory/pages/LoginPage.java
@@ -1,10 +1,9 @@
-package e2e.pages;
+package e2e.AdminHistory.pages;
 
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
 import org.openqa.selenium.support.PageFactory;
-import org.openqa.selenium.support.ui.ExpectedCondition;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.WebDriverWait;
