+++ TESTS/e2eTests/src/test/java/e2e/AdminHistory/pages/RideHistoryPage.java
@@ -1,6 +1,6 @@
-package e2e.pages;
+package e2e.AdminHistory.pages;
 
-import e2e.models.History;
+import e2e.AdminHistory.models.History;
 import org.openqa.selenium.By;
 import org.openqa.selenium.Keys;
 import org.openqa.selenium.WebDriver;
@@ -11,7 +11,6 @@ import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.Select;
 import org.openqa.selenium.support.ui.WebDriverWait;
 
-import java.security.PrivilegedAction;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
