+++ TESTS/e2eTests/src/test/java/e2e/testng.xml
@@ -3,7 +3,7 @@
     <parameter name="frontendUrl" value="http://localhost:4200"/>
     <test name="Backend E2E Tests">
         <packages>
-            <package name="e2e.tests"/>
+            <package name="e2e.AdminHistory.tests"/>
         </packages>
 
     </test>
